refactor: create profile tooltip component
